requirements.txt
used to install the required libraries for the project

!pip install requests python-dotenv

request is to call apis and 
python-dotenv is used to securely store api key

pip install colorlog